vp8_decoder: do not reset port_def in reset_stream_parameters
This is causing issues on dynamic resolution changes.
Also I found that in all other plugins it is not done. So let's
do the same here.

The exact issue was that the color format get lost because disabling
the input port caused to call this TIZ_INIT_OMX_PORT_STRUCT. Then
tiz_krn_SetParameter_internal raised an error because of 0 color format
and 0 min buffer count.